+++ CHANGELOG.rst
@@ -11,9 +11,8 @@ and this project adheres to
 --------------------
 Added
 ^^^^^
-- add Experiment class designed for mrfm experiments
+- add Experiment class designed for MRFM experiments
 - add shortcut to modify existing Models
 - add component that replaces partial parameters
 - add component and printout modifiers
-
-
+- add cli and yaml configurations